pages(register.vue)のコンポーネント分割 #17
